fix: Fix multiple iterations not working
A check using a hardcoded value was based on the input of the first
iteration. As such further iterations would fail that check causing the
wrong output to be sent back again.

Additionally a few asserts were also only based on first iteration and
had to be adjusted accordingly for multiple iterations.