refactor: Set vector capacity based on i/o size
Instead of setting it based on some hardcoded magic value.

This allows us to scale the size based on expected input.

Note that `with_capacity()` only means that the initial vector has that
size. It can still be increased over that value. It simply guarantees
that the initial vector has that size and as such won't be reallocated
when adding values up to that size.